+++ Sources/Vapor/Utilities/LifecycleHandler.swift
@@ -1,0 +1,13 @@
+import NIO
+
+public protocol LifecycleHandler {
+    func willBoot(_ application: Application) throws
+    func didBoot(_ application: Application) throws
+    func shutdown(_ application: Application)
+}
+
+extension LifecycleHandler {
+    public func willBoot(_ application: Application) throws { }
+    public func didBoot(_ application: Application) throws { }
+    public func shutdown(_ application: Application) { }
+}
